Split the html converter test and html converter style test
They were packed into a single file and it has gotten very long
as Sablon's HTML conversion support has expanded.
